+++ karen/device.py
@@ -4,6 +4,7 @@ import time
 import os
 import socket 
 import ssl
+import uuid
 from urllib.parse import urljoin
 
 from .shared import threaded, KHTTPRequestHandler, KJSONRequest, sendJSONRequest
@@ -18,7 +19,7 @@ class DeviceContainer:
     This minimizes the overhead on the device to send its collected content to the brain.
     """
     
-    def __init__(self, tcp_port=8081, hostname="localhost", ssl_cert_file=None, ssl_key_file=None, brain_url="http://localhost:8080"):
+    def __init__(self, tcp_port=8081, hostname="localhost", ssl_cert_file=None, ssl_key_file=None, brain_url="http://localhost:8080", friendlyName=None):
         """
         Device Container Initialization
         
@@ -28,6 +29,7 @@ class DeviceContainer:
             ssl_cert_file (str): The path and file name of the SSL certificate (.crt) for secure communications
             ssl_key_file (str): The path and file name of the SSL private key (.key or .pem) for secure communications
             brain_url (str):  The URL of the brain device.
+            friendlyName (str): The user-defined named for this device (a.k.a. "living room"). (optional)
         
         Both the ssl_cert_file and ssl_key_file must be present in order for SSL to be leveraged.
         """
@@ -36,7 +38,7 @@ class DeviceContainer:
         self._socket = None             # Socket object (where the listener lives)
         self._thread = None             # Thread object for TCP Server (Should be non-blocking)
         self._deviceThread = None       # Thread object for device checks (runs every 5 seconds to confirm devices are active)
-        self._isRunning = False         # Flag used to indicate if TCP server should be running
+        self.isRunning = False         # Flag used to indicate if TCP server should be running
         self._threadPool = []           # List of running threads (for incoming TCP requests)
         
         self._handlers = {}             # Handlers to be called for incoming command requests (e.g. { "KILL": kill_function }
@@ -45,6 +47,8 @@ class DeviceContainer:
         
         self.logger = logging.getLogger("CONTAINER")
         self.httplogger = logging.getLogger("HTTP")
+        
+        self.name = friendlyName
         
         # TCP Command Interface
         self.tcp_port = tcp_port if tcp_port is not None else 8080           # TCP Port for listener.
@@ -87,7 +91,7 @@ class DeviceContainer:
         ret, msg = sendJSONRequest(url, payLoad)
         return ret
     
-    def _registerWithBrain(self):
+    def registerWithBrain(self):
         """
         Sends a registration request to the brain for this client container indicating what devices it represents.
             
@@ -99,14 +103,17 @@ class DeviceContainer:
         
         data = {}
         for deviceType in self.objects:
-            friendlyNames = []
-            for item in self.objects[deviceType]:
-                if item["friendlyName"] is not None:
-                    friendlyNames.append(item["friendlyName"])
+            #friendlyNames = []
+            #for item in self.objects[deviceType]:
+            #    if item["friendlyName"] is not None:
+            #        friendlyNames.append(item["friendlyName"])
                     
-            data[deviceType] = { "count": len(self.objects[deviceType]), "names": friendlyNames }
+            #data[deviceType] = { "count": len(self.objects[deviceType]), "names": friendlyNames }
+            data[deviceType] = []
+            for i, item in enumerate(self.objects[deviceType]):
+                data[deviceType].append({ "uuid": item["uuid"], "active": item["device"].isRunning, "name": item["friendlyName"]})
             
-        result = self._sendRequest("register", { "port": self.tcp_port, "useHttp": self.use_http, "devices": data })
+        result = self._sendRequest("register", { "port": self.tcp_port, "useHttp": self.use_http, "name": self.name, "devices": data })
         if result:
             self.logger.info("Registration COMPLETE")
         else:
@@ -182,7 +189,7 @@ class DeviceContainer:
             (thread):  The thread for the container's TCP server
         """
         
-        self._isRunning = True 
+        self.isRunning = True 
                 
         self._lock.acquire()
 
@@ -200,7 +207,7 @@ class DeviceContainer:
             
         self._lock.release()
 
-        while self._isRunning:
+        while self.isRunning:
 
             try:
                 # Accept the new connection
@@ -296,11 +303,11 @@ class DeviceContainer:
             (bool):  True on success else will raise an exception.
         """
         
-        if self._isRunning:
+        if self.isRunning:
             return True 
 
         if autoRegister:
-            self._registerWithBrain()
+            self.registerWithBrain()
         
         self._thread = self._tcpServer()
         self.logger.info("Started @ "+ str(self.my_url))
@@ -308,7 +315,7 @@ class DeviceContainer:
         if autoStartDevices:
             for deviceType in self.objects:
                 for item in self.objects[deviceType]:
-                    if not item["device"]._isRunning:
+                    if not item["device"].isRunning:
                         item["device"].start()
                         
         if not useThreads:
@@ -328,16 +335,16 @@ class DeviceContainer:
             (bool):  True on success else will raise an exception.
         """
 
-        #if not self._isRunning:
+        #if not self.isRunning:
         #    return True 
                 
         if seconds > 0:
             self.logger.info("Shutting down in "+str(seconds)+" second(s).")
             for i in range(0,seconds):
-                if self._isRunning:
+                if self.isRunning:
                     time.sleep(1)
             
-            if self._isRunning and self._thread is not None:
+            if self.isRunning and self._thread is not None:
                 self.stop()
         
         
@@ -359,10 +366,10 @@ class DeviceContainer:
             (bool):  True on success else will raise an exception.
         """
         
-        if not self._isRunning:
+        if not self.isRunning:
             return True 
         
-        self._isRunning = False 
+        self.isRunning = False 
         
         i = len(self._threadPool) - 1
         while i >= 0:
@@ -412,7 +419,7 @@ class DeviceContainer:
         for ldeviceType in self.objects:
             if deviceType is None or deviceType == ldeviceType:
                 for item in self.objects[ldeviceType]:
-                    if item["device"]._isRunning:
+                    if item["device"].isRunning:
                         item["device"].stop()
             
                 if removeDevices:
@@ -420,14 +427,15 @@ class DeviceContainer:
                 
         return True
     
-    def addDevice(self, deviceType, obj, friendlyName=None, autoStart=True, autoRegister=True):
+    def addDevice(self, deviceType, obj, friendlyName=None, id=None, autoStart=True, autoRegister=True):
         """
         Adds a new input/output device and optionally starts it and registers it with the brain.
         
         Args:
             deviceType (str):  The type of the device being added (e.g. "karen.listener.Listener").
             obj (object):  The instantiated class object of the input/output device.
-            friendlyName (str):  The friendly name of the device (e.g. "living room").
+            friendlyName (str):  The friendly name of the device (e.g. "living room"). (optional)
+            id (str):  The Unique ID of this device.  If not set then will default to uuid.uuid4().
             autoStart (bool):  Indicates if this call should call the start() method on input/output device.
             autoRegister (bool):  Indicates if the client should automatically register with the brain
             
@@ -439,17 +447,17 @@ class DeviceContainer:
         if deviceType not in self.objects:
             self.objects[deviceType] = []
         
-        self.objects[deviceType].append({ "device": obj, "friendlyName": friendlyName })
+        self.objects[deviceType].append({ "device": obj, "friendlyName": friendlyName, "uuid": str(uuid.uuid4()) if id is None else str(id) })
         self.logger.info("Added " + str(deviceType))
 
         if autoStart:
-            if not obj._isRunning:
+            if not obj.isRunning:
                 self.logger.debug("Requesting start for " + str(deviceType))
                 obj.start()
                 self.logger.debug("Start request completed for " + str(deviceType))
 
-        if self._isRunning and autoRegister:
-            self._registerWithBrain()
+        if self.isRunning and autoRegister:
+            self.registerWithBrain()
             
         return True 
     
@@ -500,7 +508,9 @@ class Device:
         """
         
         self.type == "DEVICE_TYPE"
-        self._isRunning = False
+        self.uuid = str(uuid.uuid4())
+        self.name = ""
+        self.isRunning = False
         pass 
     
     def start(self, useThreads=True):
@@ -514,7 +524,7 @@ class Device:
             (bool):  True on success else will raise an exception.
         """
 
-        self._isRunning = True
+        self.isRunning = True
         return True 
     
     def wait(self, seconds=0):
@@ -540,5 +550,5 @@ class Device:
             (bool):  True on success else will raise an exception.
         """
         
-        self._isRunning = False
+        self.isRunning = False
         return True