+++ karen/handlers.py
@@ -65,9 +65,8 @@ def brain_handleImageInputData(jsonRequest):
     
     jsonRequest.container.logger.info("(" + str(jsonRequest.payload["type"]) + ") " + str(jsonRequest.payload["data"]))
     jsonRequest.container.addData(jsonRequest.payload["type"], jsonRequest.payload["data"])
-    jsonRequest.sendResponse(message="Data collected successfully.")
     
-    return True
+    return jsonRequest.sendResponse(message="Data collected successfully.")
 
 def brain_handleKillAllCommand(jsonRequest):
     """
@@ -101,12 +100,20 @@ def brain_handleRelayCommand(jsonRequest):
     my_cmd = jsonRequest.payload["command"]
     jsonRequest.container.logger.debug(my_cmd + " received.")
     
-    jsonRequest.sendResponse(False, "Command completed.") 
-    retVal = jsonRequest.container.sendRequestToDevices("control", jsonRequest.payload)
-        
+    if "data" not in jsonRequest.payload or "command" not in jsonRequest.payload["data"]:
+        return jsonRequest.sendResponse(True, "Invalid requested command.")
+    
+    inContainer = str(jsonRequest.payload["data"]["container"]) if "container" in jsonRequest.payload["data"] else None
+    inType = str(jsonRequest.payload["data"]["type"]) if "type" in jsonRequest.payload["data"] else None
+    inFilter = str(jsonRequest.payload["data"]["name"]) if "name" in jsonRequest.payload["data"] else None
+    if inFilter is None:
+        inFilter = jsonRequest.payload["data"]["id"] if "id" in jsonRequest.payload["data"] else None
+    
+    retVal = jsonRequest.container.sendRequestToDevices("control", jsonRequest.payload["data"], inContainer=inContainer, inType=inType, inFilter=inFilter)
+    
     if not retVal:
         return jsonRequest.sendResponse(True, "At least one message failed.")
-
+    
     return jsonRequest.sendResponse(False, "Command completed.")
     
 def brain_handleRelayListenerCommand(jsonRequest):
@@ -126,7 +133,7 @@ def brain_handleRelayListenerCommand(jsonRequest):
     jsonRequest.sendResponse(False, "Command completed.") 
     retVal = jsonRequest.container.sendRequestToDevices("control", jsonRequest.payload, "karen.listener.Listener")
         
-    return jsonRequest.sendResponse(False, "Command completed.") 
+    return retVal
 
 def brain_handleRelayWatcherCommand(jsonRequest):
     """
@@ -142,10 +149,17 @@ def brain_handleRelayWatcherCommand(jsonRequest):
     my_cmd = jsonRequest.payload["command"]
     jsonRequest.container.logger.debug(my_cmd + " received.")
     
-    jsonRequest.sendResponse(False, "Command completed.") 
-    retVal = jsonRequest.container.sendRequestToDevices("control", jsonRequest.payload, "karen.watcher.Watcher")
+    if my_cmd == "STOP_WATCHER":
+        my_cmd = "STOP_DEVICE"
+    elif my_cmd == "STOP_LISTENER":
+        my_cmd = "STOP_DEVICE"
         
-    return jsonRequest.sendResponse(False, "Command completed.") 
+    sendToDevice = { "command": my_cmd, "type": "karen.watcher.Watcher" }
+    
+    jsonRequest.sendResponse(False, "Command completed.") 
+    retVal = jsonRequest.container.sendRequestToDevices("control", sendToDevice, "karen.watcher.Watcher")
+        
+    return retVal
 
 def brain_handleSayData(jsonRequest):
     """
@@ -164,7 +178,7 @@ def brain_handleSayData(jsonRequest):
     
     if not jsonRequest.container.say(jsonRequest.payload["data"]):
         jsonRequest.container.logger.error("SAY command failed")
-        jsonRequest.sendResponse(True, "An error occurred")
+        return jsonRequest.sendResponse(True, "An error occurred")
         
     return jsonRequest.sendResponse(False, "Say command completed.") 
 
@@ -188,7 +202,56 @@ def device_handleStartStopListenerCommand(jsonRequest):
                 item["device"].start()
             elif my_cmd == "STOP_LISTENER":
                 item["device"].stop()
+                
+        jsonRequest.container.registerWithBrain() # Updates the status of the devices.
         
+    return jsonRequest.sendResponse(False, "Command completed.") 
+
+def device_handleStartStopCommand(jsonRequest):
+    """
+    Handles an inbound START/STOP method for a listener.  Stops the listening device but not the client container.
+    
+    Args:
+        jsonRequest (karen.shared.KJSONRequest): Object containing the inbound JSON request
+        
+    Returns:
+        (bool):  True on success and False on failure.
+    """
+
+    my_cmd = str(jsonRequest.payload["command"]).upper()
+    jsonRequest.container.logger.debug(my_cmd + " received.")
+
+    inType = str(jsonRequest.payload["type"]) if "type" in jsonRequest.payload else None
+    inName = str(jsonRequest.payload["name"]) if "name" in jsonRequest.payload else None
+    inUUID = str(jsonRequest.payload["uuid"]) if "uuid" in jsonRequest.payload else None
+
+    bChange = False
+    for deviceType in jsonRequest.container.objects:
+        for item in jsonRequest.container.objects[deviceType]:
+            if inType is None or inType == deviceType:
+                
+                bFound = False
+                if inName is None and inUUID is None:
+                    bFound = True
+                
+                elif inName is not None and item["friendlyName"] == inName:
+                    bFound = True
+                
+                elif inUUID is not None and item["uuid"] == inUUID:
+                    bFound = True
+
+                if bFound:
+                    if my_cmd == "STOP_DEVICE":
+                        item["device"].stop()
+                        bChange = True
+
+                    if my_cmd == "START_DEVICE":
+                        item["device"].start()
+                        bChange = True
+
+    if bChange:
+        ret = jsonRequest.container.registerWithBrain()
+    
     return jsonRequest.sendResponse(False, "Command completed.") 
 
 def device_handleStartStopWatcherCommand(jsonRequest):
@@ -211,6 +274,8 @@ def device_handleStartStopWatcherCommand(jsonRequest):
                 item["device"].start()
             elif my_cmd == "STOP_WATCHER":
                 item["device"].stop()
+                
+        jsonRequest.container.registerWithBrain() # Updates the status of the devices.
         
     return jsonRequest.sendResponse(False, "Command completed.")